requirements.txt
# AI Vision Dependencies
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
ultralytics>=8.0.0  # For YOLOv8
numpy>=1.24.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0  # For JWT
passlib[bcrypt]>=1.7.4  # For password hashing

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Communication
paho-mqtt>=2.0.0
websockets>=11.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0 